requirements.txt
numpy
matplotlib
jax
optax
rlax
chex
dm-haiku
tensorflow
tensorflow_probability
dm_control
dm-reverb
git@github.com:Leonardo130709/rltools.git